Test : fcm토큰 생성방식 테스트코드

+++ src/main/java/com/codenear/butterfly/notify/fcm/application/FCMTokenService.java
@@ -33,7 +33,8 @@ public class FCMTokenService {
         List<FCM> existingFcm = fcmRepository.findByToken(token);
 
         for (FCM fcm : existingFcm) {
-            fcmRepository.delete(fcm);
+            if (fcm.getMember().getId() == loginMember.getId())
+                fcmRepository.delete(fcm);
         }
 
         FCM fcm = createFCM(token, member);
